refactor: use operationsService from domains/backend in approve and transfer models
